market-analyzer-v2/requirements.txt
flask==3.0.3
flask-pymongo==3.0.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1
python-dotenv==1.0.1
bcrypt==4.2.0
pydantic==2.8.2
numpy==2.1.1
Pillow==10.4.0
tensorflow==2.20.0
opencv-python==4.10.0.84
requests==2.32.3
flask-limiter==3.7.0
Flask-SocketIO==5.3.0
python-socketio==5.9.0
python-engineio==4.7.0
keras
Werkzeug==3.0.0
simple-websocket
